Move methods out from CadRevealComposerRunner
